Update to sky news
TL;DR: Works, but not as I would like.

Sky don't provide a UUID with their RSS posts (because it's RSS)
So long story short this is currently using the title as it's ID. Now this *should* be unique, but a future version of this will use the ID of the item in the array as it's ID.

Also changed the input statement to reflect this. Might put all of this into a component?